Anatanas pointed out a couple of bugs in code
1. Application did not load in the background images for the buttons
This is done in the constructor and was part of the homework
2. The MenuState Buttons were all being drawn on top of each other in
the top left of the screen. This is because in Button::draw I forgot
to send states - the second parameter of draw and hence all of the
transforms were being neglected. Pass states to put objects where
you expect them to be!
3. The Applicatoon was not checking if the statestack was empty and
hence the Exit button callback was not exiting. Changed the run
method in Application.cpp to check if the statestack if empty and
close the window if it is.
Thanks Anatanas!