Fixed 3 major bugs that were small fixes but hard to find
TextNode text was not displaying as needed to convert hitpoints
to a string with std::to_string before rendering
Projectiles were not being rendered, or not being rendered in the
right place, This issue was more subtle. Changed the SceneNode
class so that each SceneNode has a category with a default of none
The aircraft are added to the Scene layer and commands are sent to
that layer using the categories
Missiles were not being guided. This was a simple fix to call
GuideMissiles in the World::Update()
We can now fire and enemies can fire